requirements.txt
torch==2.1.0
transformers==4.35.2
torch_geometric==2.4.0
spacy==3.7.2
pandas==2.1.3
scikit-learn==1.3.2
tqdm==4.66.1
pyyaml==6.0.1
seqeval==1.2.2
spacy-entity-linker==1.0.4